ci: move security review runner into workflow scope
