+++ .github/workflows/security-review-shared.yml
@@ -143,7 +143,7 @@ jobs:
         uses: actions/github-script@3a2844b7e9c422d3c10d287c895573f7108da1b3  # v9.0.0
         with:
           script: |
-            const runSecurityReview = require('./.github/scripts/security-review-runner.js');
+            const runSecurityReview = require('./.github/workflows/security-review-runner.js');
             await runSecurityReview({ github, context, core });
         env:
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
